refactor: use pybind11 to simplify C++ code
Use [pybind11][1] to simplify our C++ code.

The current C Python API is very verbose, as it's designed for use with
C code. However, since we're using C++, we can use something a C++ API.

[nanobind][2] is the more modern C++ Python bindling library, however,
we're **NOT** using it due to it's CMake requirement. Using CMake in a
Python extension is pretty easy if we're using [scikit-build-core][3],
but unfortuantely, `scikit-build-core` conflicts with [poetry-core][4],
which we are using for managing Python dependencies.

[1]: https://github.com/pybind/pybind11
[2]: https://github.com/wjakob/nanobind
[3]: https://github.com/scikit-build/scikit-build-core
[4]: https://github.com/python-poetry/poetry-core