feat: add propagate_const.hpp header
Taken from
https://github.com/MikeyCpp/propagate_const/blob/265598843546c16c56538590ea5d7c3186da58ec/propagate_const.hpp
under the MIT license.

Ideally we'd use [std::experimental::propagate_const][1], but it's not
available on all compilers.

I've defined the licenses using the
[REUSE Specification – Version 3.0][2]. This means that copyrighted
files have an `SPDX-License-Identifier` header, and that the licenses
are stored in a `LICENSES/` directory.

These licenses are then stored in the build Python package by using
[`tool.scikit-build.wheel.license-files`][3] to pass the licenses to
`scikit-build-core`. Unfortunately, scikit-build-core does **NOT**
support [PEP 639][].

[1]: https://en.cppreference.com/w/cpp/experimental/propagate_const
[2]: https://reuse.software/spec/
[3]: https://scikit-build-core.readthedocs.io/en/latest/configuration.html#customizing-the-built-wheel
[PEP 639]: https://peps.python.org/pep-0639