build: include and link with libirimager library
Include and link with the libirmager library.

As there may be some cases where we do not want to compile with the real
libirimager library, and we don't want to run on a real IR camera, we
can install the nqm-irimager package with
`SKBUILD_CMAKE_ARGS='-DIRImager_mock=ON' to instead use the existing
mocked code.

I've updated the CI code to automatically download the correct
libirimager version for each CPU architecture.