ci(build): use cibuildwheel version from pdm
Currently, the `build_wheels_matrix` job installs the latest version of
cibuildwheel pip package, that is [v2.15.0][1], that supports
Python 3.12.

However, the `build_wheels` job is pinned to the
`pypa/cibuildwheel/v2.14.1`, which does not support Python 3.12.

Instead, both the matrix calculation and the building should use the
same version of `cibuildwheel`, which is defined by the PDM lock file.

[1]: https://github.com/pypa/cibuildwheel/releases/tag/v2.15.0