refactor: hide `IRImager()` empty constructor
Hide the `IRImager()` empty constructor.

It's not valid to create an empty `IRImager()` object (as the camera
will have no calibration data), so we should remove this so that nobody
can use it.

I've kept it as `protected`, because we will need in the
`IRImagerMock()`, but just to create the `pImpl` object.