build(cmake): find IRImager version
Add some code in the FindIRImager.cmake script that detects the version
of the IRImager library.

Unfortuantely, there's no easy way to do this, as it's not defined in a
header anywhere. Instead, we have to compile and run a program that
calls [`evo::IRImager::getVersion()`][1] function. This can be done with
CMake's [`try_run()`][2] command.

[1]: http://documentation.evocortex.com/libirimager2/html/classevo_1_1IRImager.html#details
[2]: https://cmake.org/cmake/help/latest/command/try_run.html