feat: implement irimager code
Replace the mocked functionality in
`src/nqm/irimager/irimager_class.cpp` with real calls to the irimager
library.

As far as I can tell, the only way to get the thermal data out of the
evo::IRImager is to use `setThermalFrameCallback()`, and it will then
pick when to return the thermal data. Since callbacks aren't really a
thing in non-GUI Python code, so I've used a mutex to re-convert it
back into non-callback code.