test: add initial unit test written in C++/GTest
Add an initial basic unit test that's written in C++ using the
[GoogleTest/GTest][1] framework.

This will allow us to test C++ code directly, instead of having to go
through the Python interpreter, which will make debugging a lot easier.

These tests can be run by either
  - [`ctest`][2], or
  - pytest, using the [`pytest-cpp` package][3].

Since sckit-build-core builds everything in a temporary directory by
default, we need to set a custom `tool.scikit-build.build-dir` in our
`pyproject.toml` file so that the tests aren't automatically deleted
before they can be run.

I've also set it up so that these C++ tests are only built if you pass
`BUILD_TESTING=ON` to scikit-build-core. This is because they need
an embedded Python interpreter to run, and many servers (e.g. like the
manylinux distros in the cibuildwheeel CI action) do not support this.

[1]: https://google.github.io/googletest/
[2]: https://cmake.org/cmake/help/latest/manual/ctest.1.html
[3]: https://github.com/pytest-dev/pytest-cpp