refactor: add a `StringRingBuffer` class
Add a `StringRingBuffer` class that operates as a ring/circular buffer
for string data.

I did a quick search on GitHub for C++ ring/circular buffers, but most
of them seemed to be generic and didn't have string-specific helper
functions, which I wanted. So I just went ahead and made my own!