fix: throw if we cannot load the XML config file
Throw an `std::runtime_error` if the given XML config file passed to
the IRImager class does not exist.

Currently, reading from a non-exsistent file works,
we just get 0-length data.

Unfortunately, we can't just throw or rethrow an
[`std::ios_base::failure`][1] exception, because even though we're using
C++17, we are still using the old `_GLIBCXX_USE_CXX11_ABI=0` pre-C++11
libstdc++ ABI, therefore
[`std::io_base::failure` might not an `std::runtime_error`][2].
This means we won't be able to see the actual OS error-code that
explains why the file could not be opened.

[1]: https://en.cppreference.com/w/cpp/io/ios_base/failure
[2]: https://github.com/gcc-mirror/gcc/blob/00d16a269a9398b72b9dc9b487bbf7fecd61ffd6/libstdc%2B%2B-v3/include/bits/ios_base.h#L262-L264