build: add eigen as a dependency
Instead of working directly with `pybind11::array_t` Numpy arrays,
which leads to all sorts of issues with Python's GIL, pybind11
supports automatically converting Eigen matrices to Numpy arrays.

Eigen comes under the MPL-2.0 license, which is compatible with
the GPL-2.0-or-later license that we have to distribute nqm.irimager
under.