feat: add `get_frame_monotonic()` method
The `evo::IRDevice::getFrame()` function doesn't return a UNIX
timestamp when getting a frame.
Instead it returns a monotonic/steady timestamp.

To convert this back to a `std::chrono::system_clock`, I've added a
function called `nqm::irimager::clock_cast`, but it's imprecise,
so it's not ideal for very high frame-rates. Instead, we may want
to use a monotonic time directly.