feat: add `LoggerContextManager` Python class
Add a `LoggerContextManager` Python class that wraps the `Logger` Python
class in a context manager so that it can be used with `with:` in
Python, e.g.

```python3
from nqm.irimager import LoggerContextManager

with LoggerContextManager():
  ... # my code here
```

Unlike using `Logger` directly, using a context manager means that
cleanup works correctly, even if there is an exception.